src/pypost/util/requirements.txt
## needed package ##
numpy
scipy
simpy
h5py
pytecplot

